hub/interpreter/requirements.txt
asyncio
python-socketio
async-timeout
aiohttp
pika
requests